requirements.txt
streamlit==1.28.1
supabase==2.0.2
python-dotenv==1.0.0
pandas==2.1.3
streamlit-authenticator==0.2.3
